refactor: abi directory inside data
