+++ src/index.ts
@@ -1,7 +1,7 @@
 import setInteractionChannel from './listener/discord/setInteractionChannel';
 import setTargetChannel from './listener/discord/setTargetChannel';
 import createEtherEventListener from './scripts/createEtherEventListener';
-import QuestBoardAbi from './data/QuestBoardAbi.json';
+import QuestBoardAbi from './data/abi/QuestBoardAbi.json';
 import balancerQuestCreationListener from './listener/ethers/balancerQuestCreationListener';
 import curveQuestCreationListener from './listener/ethers/curveQuestCreationListener';
 import data from './data/data.json';
